OvmfPkg: Add a dummy FMP device
A new build flag, FMP_DUMMY_ENABLE, is added to OVMF to build the
dummy FMP device and a dummy update capsule. The dummy FMP device
does nothing but exports an entry to ESRT and the update capsule contains
no functional change. The goal is to provide the fwupd/shim developers
to verify the capsule update procedures with a VM.

The capsule, dummy-0.0.0.2.cap, is signed with the default test key and
available in Build/Ovmf${ARCH}/Capsules/TestCert_${ARCH}_DEBUG_${TOOLCHAIN}/.
The LVFS metadata and the capsule are archived in dummy-0.0.0.2.cab, so
that fwupd can install the capsule with the following command:

  $ fwupdmgr install dummy-0.0.0.2.cab

To minimize the impact of change, PEI variable and capsule processing
are not implemented. So the capsule won't be processed and the last
attempt status is always EFI_SUCCESS.

FMP_DUMMY_ENABLE **SHOULD NOT** be enabled for the production machines.
Such device is created only for some specific tests. Besides, the device
is signed with the test key in BaseTools and the private key is in public.

NOTE:
  * FmpDeviceLibDummy and GenCapsuleAll.py are borrowed from
    edk2-platforms/Vlv2TbltDevicePkg.
  * A fully functional capsule update implementation can be found in
    https://github.com/tianocore/edk2-staging/commits/Bug_1525_FmpDevicePkg_MultipleControllers_V2

Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: Ard Biesheuvel <ard.biesheuvel@arm.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Signed-off-by: Gary Lin <glin@suse.com>